Implemented Feature 9.
Application can now Gamma Encode a raw file.

A gamma encoding of 1/2.2 is assumed.

The loadRaw(string filename) takes in a raw file and reads in the float
values.
The method then converts the float to a unsigned int for colour channels r,g,b. The new gamma code is then
applied to the power of the colour channel.

unsigned int newR = MAX_RGB_VALUE * pow(r, 1.0 / 2.2);