Implemented a Emboss Filter(Feature 10).
The Emboss filter was done using a Convolution Matrix.

To get the convolved pixels, the eight pixels around the current pixel
had to be got first. The getConvolutionPixels(const int& currentPixel)
method does this.
The convolved pixels are got by multipltying each convolution Pixels r,
g and b by the embossConvolutionMatrix and setting the value
outputted equal to a int for each colour channel.

127 is then added to each colour channel to get the gryscale effect of the
emboss filter and set each int colour channel equal to the r,g,b of each pixel.

Impleented a Guassian Blur as well because it used alot of the code
already created from the emboss filter.